mind_palace/learning/session/filters.py
from django_filters import rest_framework as filters

from . import models


class LearningSessionFilterSet(filters.FilterSet):

    class Meta:
        model = models.UserLearningSession
        exclude = ('queue', 'additional_queue', 'repeated_nodes')
